Move CLient class from __main__ to __init__ script
